cache-with-redis/requirements.txt
django==3.2.4
django-redis==5.0.0
django-debug-toolbar==3.2.1